+++ core/src/main/scala/org/apache/spark/security/SSLConfig.scala
@@ -39,10 +39,8 @@ object SSLConfig extends Logging {
 
     val sparkSSLPrefix = "spark.ssl."
 
-    val vaultTrustStorePath = options.get(s"${sslType}_VAULT_TRUSTSTORE_PATH")
-    val vaultTrustStorePassPath = options.get(s"${sslType}_VAULT_TRUSTSTORE_PASS_PATH")
-    val trustStore = VaultHelper.getTrustStore(vaultTrustStorePath.get)
-    val trustPass = VaultHelper.getCertPassForAppFromVault(vaultTrustStorePassPath.get)
+    val trustStore = VaultHelper.getAllCas
+    val trustPass = VaultHelper.getCAPass
     val trustStorePath = generateTrustStore(sslType, trustStore, trustPass)
 
     logInfo(s"Setting SSL values for $sslType")
@@ -94,13 +92,15 @@ object SSLConfig extends Logging {
           s"${ConfigSecurity.secretsFolder}/key.pkcs8",
         s"$sparkSSLPrefix${sslType.toLowerCase}.caPem.path" ->
           s"${ConfigSecurity.secretsFolder}/ca.crt")
+
     trustStoreOptions ++ keyStoreOptions ++ keyPass ++ certFilesPath
   }
 
-  private def generateTrustStore(sslType: String, cas: String, password: String): String = {
+  def generateTrustStore(sslType: String, cas: String, password: String): String = {
 
     val keystore = KeyStore.getInstance("JKS")
     keystore.load(null)
+
     val certs = getBase64FromCAs(cas)
 
     certs.zipWithIndex.foreach { case (cert, index) =>
@@ -123,8 +123,8 @@ object SSLConfig extends Logging {
 
   def generatePemFile(pem: String, fileName: String): Unit = {
       formatPem(pem)
-      val bosCA = new BufferedOutputStream(new FileOutputStream(s"${ConfigSecurity.secretsFolder}" +
-        s"/$fileName"))
+      val bosCA = new BufferedOutputStream(
+        new FileOutputStream(s"${ConfigSecurity.secretsFolder}/$fileName"))
       bosCA.write(formatPem(pem).getBytes)
       bosCA.close()
     }
@@ -148,8 +148,8 @@ object SSLConfig extends Logging {
       val decrypted = pkcs8.getDecryptedBytes
       val spec = new PKCS8EncodedKeySpec(decrypted)
       val pk = KeyFactory.getInstance("RSA").generatePrivate(spec)
-      val bos = new BufferedOutputStream(new FileOutputStream(s"${ConfigSecurity.secretsFolder}" +
-        s"/key.pkcs8"))
+      val bos = new BufferedOutputStream(
+        new FileOutputStream(s"${ConfigSecurity.secretsFolder}/key.pkcs8"))
       bos.write(pk.getEncoded)
       bos.close()
     }